use float division when calcing 'bitcoin amount'

+++ app/controllers/coinbase_controller.rb
@@ -8,7 +8,7 @@ class CoinbaseController < ApplicationController
     order.price = params[:order][:total_native][:cents].to_i / 100
     order.coinbase_id = params[:order][:id]
     order.coinbase_code = params[:order][:button][:id]
-    order.bitcoin_amount = params[:order][:total_btc][:cents].to_i / 100000000
+    order.bitcoin_amount = params[:order][:total_btc][:cents].to_f / 100000000.0
     order.bitcoin_payout_address = orderable.user.bitcoin_address
     order.finish!
     render nothing: true, status: 200
